Moved the Admin and User DashBoard component to a more appropriate file
